refactor(platform-api): extract shared constants into constants.py
