+++ platform/api/hexgate_api/demo.py
@@ -60,7 +60,7 @@ def enable_demo(app: FastAPI) -> None:
         from hexgate_api.auth import cookie_transport, get_jwt_strategy
         from hexgate_api.core.db import async_session_factory
         from hexgate_api.models import User
-        from hexgate_api.services import DEFAULT_USER_ID
+        from hexgate_api.constants import DEFAULT_USER_ID
 
         async with async_session_factory() as session:
             user = await session.get(User, DEFAULT_USER_ID)
